追加実装

+++ app/Http/Controllers/PostController.php
@@ -63,6 +63,9 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if (Auth::id() !== $post->user_id) {
+            return abort(404);
+        }
         $post->delete();
 
         return redirect()->route('posts.index');
